move security controller to utils
